Fix: Memory Leak
